Move the log-stream tests into the existing test file
Requested in review: the new tests belong in server/logging/log_test.go
rather than a separate close_test.go. Also shortens the comment in the
done handler.
